+++ api-gateway/src/main/java/com/petclinic/bffapigateway/domainclientlayer/BillServiceClient.java
@@ -1,8 +1,6 @@
 package com.petclinic.bffapigateway.domainclientlayer;
 
-import com.petclinic.bffapigateway.dtos.BillDetails;
-import com.petclinic.bffapigateway.dtos.OwnerDetails;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.petclinic.bffapigateway.dtos.Bills.BillDetails;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
