+++ api-gateway/src/main/java/com/petclinic/bffapigateway/domainclientlayer/InventoryServiceClient.java
@@ -1,14 +1,12 @@
 
 package com.petclinic.bffapigateway.domainclientlayer;
 
-import com.petclinic.bffapigateway.dtos.BundleDetails;
-import com.petclinic.bffapigateway.dtos.ProductRequestDTO;
-import com.petclinic.bffapigateway.dtos.ProductResponseDTO;
+import com.petclinic.bffapigateway.dtos.Inventory.ProductRequestDTO;
+import com.petclinic.bffapigateway.dtos.Inventory.ProductResponseDTO;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
