+++ api-gateway/src/main/java/com/petclinic/bffapigateway/dtos/Auth/Register.java
@@ -1,26 +1,25 @@
-package com.petclinic.bffapigateway.dtos;
+package com.petclinic.bffapigateway.dtos.Auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.Set;
+/**
+ * Created by IntelliJ IDEA.
+ *
+ * User: @Fube
+ * Date: 2021-10-15
+ * Ticket: feat(APIG-CPC-354)
+ */
 
 @Data
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 @Builder(toBuilder = true)
-public class UserDetails {
-
-    private long id;
-
-    private String username;
-
-    private String password;
+public class Register {
 
     private String email;
-
-    private Set<Role> roles;
+    private String username;
+    private String password;
 }
-
