+++ api-gateway/src/main/java/com/petclinic/bffapigateway/dtos/Auth/RoleIdLess.java
@@ -1,4 +1,4 @@
-package com.petclinic.bffapigateway.dtos;
+package com.petclinic.bffapigateway.dtos.Auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
