--- api-gateway/src/main/java/com/petclinic/bffapigateway/dtos/BundleDetails.java
@@ -1,14 +1,0 @@
-
-package com.petclinic.bffapigateway.dtos;
-
-import lombok.Data;
-
-import java.util.Date;
-
-@Data
-public class BundleDetails {
-    private String bundleUUID;
-    private String item;
-    private int quantity;
-    private Date expiryDate;
-}
