+++ api-gateway/src/main/java/com/petclinic/bffapigateway/dtos/Inventory/ProductRequestDTO.java
@@ -1,4 +1,4 @@
-package com.petclinic.bffapigateway.dtos;
+package com.petclinic.bffapigateway.dtos.Inventory;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,10 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class ProductResponseDTO {
-    private String id;
-    private String productId;
-    private String inventoryId;
+public class ProductRequestDTO {
     private String productName;
     private String productDescription;
     private Double productPrice;
