+++ api-gateway/src/main/java/com/petclinic/bffapigateway/dtos/Pets/PetDetails.java
@@ -1,5 +1,6 @@
-package com.petclinic.bffapigateway.dtos;
+package com.petclinic.bffapigateway.dtos.Pets;
 
+import com.petclinic.bffapigateway.dtos.Vets.VisitDetails;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
