+++ api-gateway/src/main/java/com/petclinic/bffapigateway/dtos/Pets/PetType.java
@@ -1,4 +1,4 @@
-package com.petclinic.bffapigateway.dtos;
+package com.petclinic.bffapigateway.dtos.Pets;
 
 import lombok.Data;
