+++ api-gateway/src/main/java/com/petclinic/bffapigateway/dtos/Visits/Visits.java
@@ -1,5 +1,6 @@
-package com.petclinic.bffapigateway.dtos;
+package com.petclinic.bffapigateway.dtos.Visits;
 
+import com.petclinic.bffapigateway.dtos.Vets.VisitDetails;
 import lombok.Value;
 
 import java.util.ArrayList;
