+++ auth-service/src/main/java/com/auth/authservice/presentationlayer/User/UserIDLessRoleLessDTO.java
@@ -20,9 +20,6 @@
 
 package com.auth.authservice.presentationlayer.User;
 
-import com.auth.authservice.Util.Configuration.PasswordStrengthCheck;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotEmpty;
 import lombok.*;
