+++ auth-service/src/main/java/com/auth/authservice/presentationlayer/User/UserIDLessUsernameLessDTO.java
@@ -1,10 +1,6 @@
 package com.auth.authservice.presentationlayer.User;
 
 
-import com.auth.authservice.Util.Configuration.PasswordStrengthCheck;
-import jakarta.persistence.Column;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotEmpty;
 import lombok.*;
 
 @Getter
